change in script again1

+++ Jenkinsfile
@@ -40,8 +40,11 @@ pipeline {
         
         stage('Docker deploy'){
             steps {
-            	bat 'docker stop api-demo'
-               	bat 'docker rm api-demo'
+            	 def imageExists = bat(script: "docker images -q api-demo", returnStdout: true) == 0
+    			if(imageExists){
+        			   	bat 'docker stop api-demo'
+               			bat 'docker rm api-demo'	
+    			}
                 bat 'docker run --name api-demo -itd -p  8085:8085 sanjay872/docker_jenkins_springboot:%BUILD_NUMBER%'
             }
         }
